// src/app/core/models/questionnaire-report.ts
export class QuestionnaireReport {
    nameInIC: string;
    preferredName: string;
    email: string;
    phoneNumber: string;
    answer1: string = '';
    answer2: string = '';
    answer3: string = '';
    answer4: string = '';
    answer5: string = '';
    answer6: string = '';
    answer7: string = '';
    answer8: string = '';
    answer9: string = '';
    answer10: string = '';
    answer11: string = '';
    answer12: string = '';
    answer13: string = '';
    answer14: string = '';
}